requirements.txt
django-mptt
django-authority
feincms
django-admin-uploads
